Sonar Avoid Logic
I added what I think will work for the sonar avoid logic.  If there is a
positive distance, it triggers the behavior.  Then, the percentage of
the range is calculated (up to 300 cm range).  The turning speed fo the
wheels is then proportionate to the closeness to the object.  Needs to
be tested.