Turning Speed Change
In a nutshell, I made the speed propotional to the distance away by
adding the turning constant.  We can adjust that constant to adjust the
speed.

I added a turning_constant so that instead of increasing the wheel speed
(from a base of 200) by the percent of the distance range (0-300cm -->
0-100%), it multiplies that distance percentage by the constant (I tried
0.5) so it isn't so drastic.  We can try tweaking that constant from now
on.